Switch image library (#25)
* include stb image/image_write

* remove Boost and libPNG dependencies by switching to stb

* remove boost/libpng remnants from cmake files

* set compiler options so mac/linux are equally oversensitive to signedness as Windows

* license details

* fix signed/unsigned type issues

* additional flags

* add include required on windows

* prefer static casts to C-style

* check not on windows before setting gcc/clang options

+++ test/helpers/helpers.h
@@ -13,7 +13,7 @@ inline fs::path get_temp_filepath(const std::string &extension, const std::strin
         {
             filename << label << "-";
         }
-        std::srand(std::time({}));
+        std::srand(static_cast<unsigned int>(std::time({})));
         filename << std::rand() << "." << extension;
         auto filepath = fs::path(fs::temp_directory_path() / filename.str());
